+++ src/main.rs
@@ -1,4 +1,5 @@
 mod audio;
+mod diagnostics;
 mod entities;
 mod mechanics;
 mod plugins;
@@ -7,14 +8,20 @@ mod visuals;
 use bevy::prelude::*;
 use bevy_ecs_ldtk::prelude::*;
 use bevy_kira_audio::AudioPlugin;
+use plugins::developer_helpers::SmartAssetIoPlugin;
 
 fn main() {
     App::new()
-        .add_plugins(DefaultPlugins.set(AssetPlugin {
-            //For dev purposes only. REMOVE WHEN GIVING TO PLAYERS!
-            watch_for_changes: true,
-            ..default()
-        }))
+        .add_plugins(
+            DefaultPlugins
+                .set(AssetPlugin {
+                    //For dev purposes only. REMOVE WHEN GIVING TO PLAYERS!
+                    watch_for_changes: true,
+                    ..default()
+                })
+                .build()
+                .add_before::<bevy::asset::AssetPlugin, _>(SmartAssetIoPlugin),
+        )
         .add_plugin(LdtkPlugin)
         .add_plugin(AudioPlugin)
         .add_plugin(plugins::levels::LevelsPlugin)