+++ src/plugins/mod.rs
@@ -1,3 +1,4 @@
+pub mod developer_helpers;
 pub mod levels;
 pub mod main_menu;
 pub mod playable_character;